Update Jenikins Configg

+++ Jenkinsfile
@@ -11,16 +11,29 @@ pipeline {
 
   stages {
     stage('Stop Containers and Remove Image') {
-      steps {
-        sh 'docker stop cce-website-web-1'
-        sh 'docker rm cce-website-web-1'
-        sh 'docker rmi -f  cce-website-web'
-      }
+    steps {
+        script {
+            /* groovylint-disable-next-line LineLength */
+            def containerStatus = sh(returnStatus: true, script: 'docker inspect -f {{.State.Status}} cce-website-web-1').trim()
+            /* groovylint-disable-next-line NestedBlockDepth */
+            if (containerStatus == 'running') {
+            sh 'docker stop cce-website-web-1'
+            }
+            sh 'docker rm -f cce-website-web-1'
+            
+            /* groovylint-disable-next-line LineLength */
+            def imageExists = sh(returnStatus: true, script: 'docker image inspect cce-website-web > /dev/null 2>&1').trim()
+            /* groovylint-disable-next-line NestedBlockDepth */
+            if (imageExists == 0) {
+            sh 'docker rmi -f cce-website-web'
+            }
+        }
+    }
     }
 
       stage('Pull from Git') {
       steps {
-        sh "cd cce-website  && git pull origin production"
+        sh "cd ~/cce-website  && git pull origin production"
       }
       }
     stage('Deploy') {
